requirements.txt
django==2.2.1
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.3
markdown==3.1
Werkzeug==0.15.2
psycopg2-binary
pytest
pytest-django
pytest-sugar
black
flake8
